requirements.txt
img2pdf>=0.4.4
Pillow>=10.0.0